3_batch_pred_sproc/requirements.txt
snowflake-snowpark-python
snowflake-ml-python
toml # for local development
pandas
cachetools
